metal: measure system-wide wired memory in test
